0dqub1-django-blog-api-n-1-query-performance-fix/requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14.0
django-filter>=23.0
pytest>=8.0.0
pytest-django>=4.8.0
pytest-json-report>=1.5.0
